+++ packages/blocks/core/js/wordpress/post-title/index.js
@@ -15,7 +15,13 @@ export const PostTitle: BlockType = {
 	name: 'blockeraPostTitle',
 	targetBlock: 'core/post-title',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;